Move circuit.hpp into separate directory for improved organization
