+++ example/quantum_phase_estimation_diagonal.cpp
@@ -10,7 +10,7 @@
 #include <mini-qiskit/decomposed/build_decomposed_circuit.hpp>
 #include <mini-qiskit/circuit_operations/make_binary_controlled_circuit.hpp>
 #include <mini-qiskit/gates/fourier.hpp>
-#include <mini-qiskit/circuit.hpp>
+#include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
 #include <mini-qiskit/state.hpp>
