+++ example/quantum_phase_estimation_trotter20.cpp
@@ -11,7 +11,7 @@
 #include <mini-qiskit/decomposed/build_decomposed_circuit.hpp>
 #include <mini-qiskit/circuit_operations/make_binary_controlled_circuit.hpp>
 #include <mini-qiskit/gates/fourier.hpp>
-#include <mini-qiskit/circuit.hpp>
+#include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
 #include <mini-qiskit/state.hpp>
 #include <mini-qiskit/io/statevector.hpp>
