+++ example/read_decomposition_file.cpp
@@ -6,7 +6,7 @@
 #include <mini-qiskit/decomposed/read_decomposition_file.hpp>
 #include <mini-qiskit/decomposed/build_decomposed_circuit.hpp>
 #include <mini-qiskit/common/print.hpp>
-#include <mini-qiskit/circuit.hpp>
+#include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
 #include <mini-qiskit/state.hpp>
