+++ example/save_statevector_example.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <sstream>
 
-#include <mini-qiskit/circuit.hpp>
+#include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
 #include <mini-qiskit/state.hpp>
 #include <mini-qiskit/io/statevector.hpp>
