+++ include/mini-qiskit/circuit/circuit.hpp
@@ -15,7 +15,7 @@
 #include "mini-qiskit/common/utils.hpp"
 #include "mini-qiskit/gates/common_u_gates.hpp"
 #include "mini-qiskit/primitive_gate.hpp"
-#include "mini-qiskit/classical_register.hpp"
+#include "mini-qiskit/circuit/classical_register.hpp"
 
 namespace mqis
 {
@@ -365,15 +365,15 @@ public:
         return unitary_gates_[matrix_index];
     }
 
-    auto control_flow_circuit(std::size_t index) const noexcept -> const std::unique_ptr<QuantumCircuit>&
-    {
-        return control_flow_circuits_[index];
-    }
+    // auto control_flow_circuit(std::size_t index) const noexcept -> const std::unique_ptr<QuantumCircuit>&
+    // {
+    //     return control_flow_circuits_[index];
+    // }
 
-    auto control_flow_predicate(std::size_t index) const noexcept -> const std::function<int(const ClassicalRegister&)>
-    {
-        return control_flow_predicates_[index];
-    }
+    // auto control_flow_predicate(std::size_t index) const noexcept -> const std::function<int(const ClassicalRegister&)>
+    // {
+    //     return control_flow_predicates_[index];
+    // }
 
     friend auto append_circuits(QuantumCircuit left, const QuantumCircuit& right) -> QuantumCircuit;
     friend void extend_circuit(QuantumCircuit& left, const QuantumCircuit& right);
@@ -387,8 +387,8 @@ private:
     // IDEA: instead of splitting all of this up, create a single ControlFlowObject that holds
     // a single quantum circuit, a std::optional<> of another quantum circuit, and the control
     // flow predicate; it will save a lot of bookkeeping;
-    std::vector<std::unique_ptr<QuantumCircuit>> control_flow_circuits_ {};
-    std::vector<std::function<int(const ClassicalRegister&)>> control_flow_predicates_ {};
+    // std::vector<std::unique_ptr<QuantumCircuit>> control_flow_circuits_ {};
+    // std::vector<std::function<int(const ClassicalRegister&)>> control_flow_predicates_ {};
 
     void check_qubit_range_(std::size_t target_index, std::string_view qubit_name, std::string_view gate_name)
     {
