+++ include/mini-qiskit/circuit_operations/append_circuits.hpp
@@ -5,7 +5,7 @@
 #include <stdexcept>
 
 #include "mini-qiskit/primitive_gate.hpp"
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 
 namespace impl_mqis
 {
