+++ include/mini-qiskit/circuit_operations/compare_circuits.hpp
@@ -6,7 +6,7 @@
 #include <stdexcept>
 #include <vector>
 
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/primitive_gate.hpp"
 #include "mini-qiskit/common/matrix2x2.hpp"
 #include "mini-qiskit/gates/common_u_gates.hpp"
