+++ include/mini-qiskit/circuit_operations/make_binary_controlled_circuit.hpp
@@ -3,7 +3,7 @@
 #include <stdexcept>
 #include <vector>
 
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/common/utils.hpp"
 #include "mini-qiskit/gates/multiplicity_controlled_u_gate.hpp"
 #include "mini-qiskit/circuit_operations/append_circuits.hpp"
