+++ include/mini-qiskit/circuit_operations/make_controlled_circuit.hpp
@@ -4,7 +4,7 @@
 #include <stdexcept>
 #include <unordered_set>
 
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/common/utils.hpp"
 #include "mini-qiskit/gates/common_u_gates.hpp"
 #include "mini-qiskit/gates/multiplicity_controlled_u_gate.hpp"
