+++ include/mini-qiskit/decomposed/read_tangelo_file.hpp
@@ -7,7 +7,7 @@
 #include <string>
 
 #include <mini-qiskit/primitive_gate.hpp>
-#include <mini-qiskit/circuit.hpp>
+#include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/common/utils.hpp>
 #include <mini-qiskit/gates/swap.hpp>
