+++ include/mini-qiskit/mini-qiskit.hpp
@@ -2,7 +2,7 @@
 
 #include <mini-qiskit/calculations/measurements.hpp>
 #include <mini-qiskit/calculations/probabilities.hpp>
-#include <mini-qiskit/circuit.hpp>
+#include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/primitive_gate.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
 #include <mini-qiskit/state.hpp>
