+++ qpe_tangelo_project/qpe_tangelo_simulate_minimal.cpp
@@ -7,7 +7,7 @@
 #include <mini-qiskit/decomposed/read_tangelo_file.hpp>
 #include <mini-qiskit/circuit_operations/append_circuits.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
-#include <mini-qiskit/circuit.hpp>
+#include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/state.hpp>
 #include <mini-qiskit/io/statevector.hpp>
