+++ test/source/calculations/probabilities_test.cpp
@@ -10,7 +10,7 @@
 #include <catch2/matchers/catch_matchers_vector.hpp>
 
 #include "mini-qiskit/calculations/probabilities.hpp"
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/simulation/simulate.hpp"
 #include "mini-qiskit/state.hpp"
