+++ test/source/circuit_operations/make_binary_controlled_circuit_test.cpp
@@ -2,7 +2,7 @@
 #include <catch2/generators/catch_generators.hpp>
 #include <catch2/generators/catch_generators_range.hpp>
 
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/simulation/simulate.hpp"
 #include "mini-qiskit/state.hpp"
 #include "mini-qiskit/gates/common_u_gates.hpp"
