+++ test/source/circuit_operations/make_controlled_circuit_test.cpp
@@ -1,7 +1,7 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/generators/catch_generators.hpp>
 
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/state.hpp"
 #include "mini-qiskit/simulation/simulate.hpp"
 #include "mini-qiskit/gates/common_u_gates.hpp"
