+++ test/source/common/append_circuits_test.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch_test_macros.hpp>
 
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/gates/common_u_gates.hpp"
 #include "mini-qiskit/circuit_operations/append_circuits.hpp"
 #include "mini-qiskit/circuit_operations/compare_circuits.hpp"
