+++ test/source/decomposed/build_decomposed_circuit_test.cpp
@@ -11,7 +11,7 @@
 #include <mini-qiskit/decomposed/decomposed_gate.hpp>
 #include <mini-qiskit/decomposed/read_decomposition_file.hpp>
 #include <mini-qiskit/decomposed/build_decomposed_circuit.hpp>
-#include <mini-qiskit/circuit.hpp>
+#include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
 #include <mini-qiskit/state.hpp>
