+++ test/source/gates/control_swap_test.cpp
@@ -3,7 +3,7 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/generators/catch_generators.hpp>
 
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/state.hpp"
 #include "mini-qiskit/simulation/simulate.hpp"
 #include "mini-qiskit/gates/swap.hpp"
