+++ test/source/gates/fourier_test.cpp
@@ -6,7 +6,7 @@
 #include <catch2/generators/catch_generators.hpp>
 #include <catch2/matchers/catch_matchers_floating_point.hpp>
 
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/common/mathtools.hpp"
 #include "mini-qiskit/gates/fourier.hpp"
 #include "mini-qiskit/simulation/simulate.hpp"
