+++ test/source/gates/toffoli_test.cpp
@@ -4,7 +4,7 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/generators/catch_generators.hpp>
 
-#include "mini-qiskit/circuit.hpp"
+#include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/gates/toffoli.hpp"
 #include "mini-qiskit/simulation/simulate.hpp"
 #include "mini-qiskit/state.hpp"
