+++ test/source/simulation/measure_test.cpp
@@ -9,7 +9,7 @@
 #include <catch2/generators/catch_generators.hpp>
 
 #include <mini-qiskit/common/print.hpp>
-#include <mini-qiskit/circuit.hpp>
+#include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/state.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
 #include <mini-qiskit/simulation/measure.hpp>
